Add icy radio metadta - fallback to old behavior without metadta, use hook
