+++ src/app/components/player/radio-info.tsx
@@ -1,59 +1,13 @@
 import { RadioIcon } from 'lucide-react'
-import { Fragment, useEffect, useState } from 'react'
+import { Fragment } from 'react'
 import { useTranslation } from 'react-i18next'
 import { MarqueeTitle } from "@/app/components/fullscreen/marquee-title.tsx"
 import { Radio } from '@/types/responses/radios'
-import IcecastMetadataStats from 'icecast-metadata-stats'
+import { useRadioMetadata } from '../../hooks/use-radio-metadata'
 
 export function RadioInfo({ radio }: { radio: Radio | undefined }) {
   const { t } = useTranslation()
-  const [radioMetadata , setRadioMetadata] = useState({title: "", artist: ""})
-  const streamUrl = radio?.streamUrl
-
-  useEffect(() => {
-    if (!streamUrl) {
-      setRadioMetadata({artist: "", title: ""})
-      return
-    }
-    let iceListener: IcecastMetadataStats
-    let prevStreamTitle = ""
-
-    try {
-      iceListener = new IcecastMetadataStats(streamUrl, {
-        interval: 10,
-        onStats: (stats) => {
-          const streamTitle = stats.StreamTitle ? stats.StreamTitle : stats.icy?.StreamTitle
-          if (streamTitle && prevStreamTitle !== streamTitle) {
-            prevStreamTitle = streamTitle
-            const i = streamTitle.indexOf(' - ')
-
-            if (i < 0) {
-              setRadioMetadata({title: streamTitle.trim(), artist: ""})
-            } else {
-              const trackArtist = streamTitle.slice(0, i).trim()
-              const trackTitle = streamTitle.slice(i + 3).trim()
-              setRadioMetadata({title: trackTitle, artist: trackArtist})
-            }
-          } else if (prevStreamTitle !== streamTitle) {
-            prevStreamTitle = streamTitle
-            setRadioMetadata({title: "", artist: ""})
-          }
-        },
-        sources: ['icy'],
-      })
-
-      iceListener.start()
-    } catch {
-      setRadioMetadata({title: "", artist: ""})
-    }
-
-    return () => {
-      if (iceListener) {
-        iceListener.stop()
-      }
-      setRadioMetadata({title: "", artist: ""})
-    }
-  }, [streamUrl])
+  const radioMetadata = useRadioMetadata(radio)
 
   return (
     <Fragment>
@@ -66,25 +20,39 @@ export function RadioInfo({ radio }: { radio: Radio | undefined }) {
       </div>
       <div className="flex flex-col w-[66%] max-w-full justify-end text-left overflow-hidden">
         {radio ? (
-          <Fragment>
-            <MarqueeTitle gap="mr-6" >
-              <span className="text-sm font-medium" data-testid="radio-title">
-                { radioMetadata.title !== "" ? radioMetadata.title : "—" }
+          radioMetadata.title !== "" ? (
+            <Fragment>
+              <MarqueeTitle gap="mr-6" >
+                <span className="text-sm font-medium" data-testid="radio-title">
+                  { radioMetadata.title }
+                </span>
+              </MarqueeTitle>
+              <span
+                className="text-xs font-light text-muted-foreground"
+                data-testid="radio-artist"
+              >
+                { radioMetadata.artist }
               </span>
-            </MarqueeTitle>
-            <span
-              className="text-xs font-light text-muted-foreground"
-              data-testid="radio-artist"
-            >
-              { radioMetadata.artist }
-            </span>
-            <span
-              className="text-xs font-light text-muted-foreground"
-              data-testid="radio-name"
-            >
-              { radio.name }
-            </span>
-          </Fragment>
+              <span
+                className="text-xs font-light text-muted-foreground"
+                data-testid="radio-name"
+              >
+                { radio.name }
+              </span>
+            </Fragment>
+          ) : (
+            <Fragment>
+              <span className="text-sm font-medium" data-testid="radio-name">
+                { radio.name }
+              </span>
+                <span
+                  className="text-xs font-light text-muted-foreground"
+                  data-testid="radio-label"
+                >
+                { t('radios.label') }
+              </span>
+            </Fragment>
+          )
         ) : (
           <span className="text-sm font-medium" data-testid="radio-no-playing">
             {t('player.noRadioPlaying')}
